Move LogicUtils to its own file
